update
ça bouge pas trop mais on travail hein

+++ database/migrations/2020_10_05_115626_create_comments_table.php
@@ -15,8 +15,8 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('task_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('task_id');
             $table->string('text');
             $table->timestamps();
         });
